Moved RealtimeSegmentDataManager to be in realtime package rather than offline
